Need to use `%{%...%}` instead of `%!...` for correct file name
Using %! resulted in the same filename (the filename of the current
buffer) being displayed in all statuslines and winbars. Not what we want
at all. We want the name of the actual buffer (usually different)
to display in each statusline and winbar.

Using %{%...%} fixed that issue; I believe it currectly evaluates the
value of the current buffer in each buffer context.

This post discusses the issue: https://is.gd/89EH4R