New statusline segment, active buffer-attached LSP client names
This new segment, enabled by default, is displayed after the plugins
notification segment; but will only be displayed if the statusline is at
least 120 columns wide.

An option is provided to disable this new segment if desired.

Note, eventually I also want to add in LSP progress messages as well
(likely in the center position).