fix: update statusline upon Gitsigns events
The statusline was NOT updating immediately when Gitsigns changes were
occurring (only after cursor move). Turns out this used to work in
Neovim 0.8, but not since then.

Gitsigns provides an autocmd to notify when changes have occurred.
Hooking into that now results in immediate updates to the statusline for any
Git changes, even ones that occur outside Neovim itself such as a branch
change.